requirements.txt
amqp==5.0.6
appdirs==1.4.3
asgiref==3.3.4
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
bootstrap-admin==0.4.4
CacheControl==0.12.6
celery==5.0.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
contextlib2==0.6.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.0
distro==1.4.0
Django==3.2.3
django-filter==2.4.0
django-queryset-csv==1.1.0
django-redis==4.12.1
django-rest-swagger==2.2.0
django-tabbed-admin==1.0.4
djangorestframework==3.11.0
djangorestframework-csv==2.1.1
drf-yasg==1.17.1
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
importlib-metadata==4.0.1
inflection==0.5.1
ipaddr==2.2.0
itypes==1.2.0
Jinja2==2.11.3
jsonschema==3.2.0
kombu==5.0.2
lockfile==0.12.2
loguru==0.5.3
MarkupSafe==1.1.1
msgpack==0.6.2
openapi-codec==1.3.2
packaging==20.3
pep517==0.8.2
progress==1.5
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
pycodestyle==2.7.0
pyparsing==2.4.6
pyrsistent==0.17.3
python-dotenv==0.17.1
pytoml==0.1.21
pytz==2021.1
PyYAML==5.4.1
redis==3.5.3
requests==2.22.0
retrying==1.3.3
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.14.0
sqlparse==0.4.1
swagger-spec-validator==2.7.3
toml==0.10.2
typing-extensions==3.10.0.0
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.8
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
zipp==3.4.1
